+++ src/Channel/operator_cmd/INVITE.cpp
@@ -58,23 +58,3 @@ void invite(Client& sender, const Command& cmd, ClientManager &clientManager, Ch
 	Channel& channel = channelManager.get_channel(channel_name);
 	channel.invite(sender, target_user);
 }
-
-static bool require_not_member(Client& sender, Client& target, Channel& channel)
-{
-	if (channel.is_member(target))
-	{
-		send_errmsg(sender, 443, target.get_nick()+ " "+channel.get_name() +" :is already on channel");
-		return false;
-	}
-	return true;
-}
-
-void Channel::invite(Client &sender, Client& target)
-{
-	if (!require_authed(sender)) return;
-	if (!require_nick_user(sender)) return;
-	if (!require_operator(sender)) return;
-	if (!require_not_member(sender, target, *this)) return;
-	invited.insert(target);
-	send_msg(target, sender.get_nick() + " invites you to join " + get_name());
-}
